Feat: 관리자 리뷰 삭제 기능 추가

+++ src/review/router.js
@@ -164,24 +164,40 @@ router.patch(
   })
 );
 
+// 리뷰 삭제
 router.delete(
   "/:reviewId",
   JwtMiddleware.checkToken,
   asyncHandler(async (req, res) => {
-    const userId = req.token.userId;
-    const reviewId = req.params.reviewId;
+    try {
+      const userId = req.token.userId;
+      const reviewId = req.params.reviewId;
 
-    const deletedReview = await ReviewService.deleteReview(userId, reviewId);
-    if (deletedReview) {
-      res.status(200).json({
-        status: 200,
-        message: "게시글이 삭제되었습니다.",
-        data: deletedReview,
-      });
-    } else {
-      res.status(404).json({
-        status: 404,
-        message: "게시글이 존재하지 않습니다.",
+      const isAdmin = req.token.role === "admin";
+      if (!isAdmin && userId !== req.token.userId) {
+        return res.status(403).json({
+          status: 403,
+          message: "사용자 권한이 없습니다.",
+        });
+      }
+
+      const deletedReview = await ReviewService.deleteReview(userId, reviewId);
+
+      if (deletedReview) {
+        res.status(200).json({
+          data: deletedReview,
+        });
+      } else {
+        res.status(404).json({
+          data: deletedReview,
+        });
+      }
+    } catch (error) {
+      console.error(error);
+      res.status(500).json({
+        status: 500,
+        message: "Internal Server Error",
+        error: error.message,
       });
     }
   })
